contract_bindings_alloy::feecontract
Module FeeContract

Generated by the following Solidity interface…

interface FeeContract {
    error AddressEmptyCode(address target);
    error DepositTooLarge();
    error DepositTooSmall();
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error FailedInnerCall();
    error FunctionDoesNotExist();
    error InvalidInitialization();
    error InvalidUserAddress();
    error NoFunctionCalled();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error UUPSUnauthorizedCallContext();
    error UUPSUnsupportedProxiableUUID(bytes32 slot);

    event Deposit(address indexed user, uint256 amount);
    event Initialized(uint64 version);
    event Log(string func, uint256 gas);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event Upgrade(address implementation);
    event Upgraded(address indexed implementation);

    constructor();

    fallback() external payable;

    receive() external payable;

    function UPGRADE_INTERFACE_VERSION() external view returns (string memory);
    function balances(address user) external view returns (uint256 amount);
    function deposit(address user) external payable;
    function getVersion() external pure returns (uint8 majorVersion, uint8 minorVersion, uint8 patchVersion);
    function initialize(address multisig) external;
    function maxDepositAmount() external view returns (uint256);
    function minDepositAmount() external view returns (uint256);
    function owner() external view returns (address);
    function proxiableUUID() external view returns (bytes32);
    function renounceOwnership() external;
    function transferOwnership(address newOwner) external;
    function upgradeToAndCall(address newImplementation, bytes memory data) external payable;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "fallback",
    "stateMutability": "payable"
  },
  {
    "type": "receive",
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "UPGRADE_INTERFACE_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "balances",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "deposit",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "getVersion",
    "inputs": [],
    "outputs": [
      {
        "name": "majorVersion",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "minorVersion",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "patchVersion",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "multisig",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "maxDepositAmount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "minDepositAmount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "proxiableUUID",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "upgradeToAndCall",
    "inputs": [
      {
        "name": "newImplementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "event",
    "name": "Deposit",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Log",
    "inputs": [
      {
        "name": "func",
        "type": "string",
        "indexed": false,
        "internalType": "string"
      },
      {
        "name": "gas",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgrade",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "DepositTooLarge",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DepositTooSmall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedInnerCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FunctionDoesNotExist",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidUserAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoFunctionCalled",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "UUPSUnauthorizedCallContext",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnsupportedProxiableUUID",
    "inputs": [
      {
        "name": "slot",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315. solidity error AddressEmptyCode(address target);

Deposit

Event with signature Deposit(address,uint256) and selector 0xe1fffcc4923d04b559f4d29a8bfc6cda04eb5b0d3c460751c2402c5c5cc9109c. solidity event Deposit(address indexed user, uint256 amount);

DepositTooLarge

Custom error with signature DepositTooLarge() and selector 0xc56d46d3. solidity error DepositTooLarge();

DepositTooSmall

Custom error with signature DepositTooSmall() and selector 0x6ba4a1c7. solidity error DepositTooSmall();

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3. solidity error ERC1967InvalidImplementation(address implementation);

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f. solidity error ERC1967NonPayable();

FailedInnerCall

Custom error with signature FailedInnerCall() and selector 0x1425ea42. solidity error FailedInnerCall();

FeeContractInstance

A FeeContract instance.

FunctionDoesNotExist

Custom error with signature FunctionDoesNotExist() and selector 0xa9ad62f8. solidity error FunctionDoesNotExist();

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2. solidity event Initialized(uint64 version);

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9. solidity error InvalidInitialization();

InvalidUserAddress

Custom error with signature InvalidUserAddress() and selector 0x702b3d90. solidity error InvalidUserAddress();

Log

Event with signature Log(string,uint256) and selector 0xdd970dd9b5bfe707922155b058a407655cb18288b807e2216442bca8ad83d6b5. solidity event Log(string func, uint256 gas);

NoFunctionCalled

Custom error with signature NoFunctionCalled() and selector 0xbc8eca1b. solidity error NoFunctionCalled();

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8. solidity error NotInitializing();

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7. solidity error OwnableInvalidOwner(address owner);

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7. solidity error OwnableUnauthorizedAccount(address account);

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0. solidity event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);

UPGRADE_INTERFACE_VERSIONCall

Function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc. solidity function UPGRADE_INTERFACE_VERSION() external view returns (string memory);

UPGRADE_INTERFACE_VERSIONReturn

Container type for the return parameters of the UPGRADE_INTERFACE_VERSION() function.

UUPSUnauthorizedCallContext

Custom error with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba. solidity error UUPSUnauthorizedCallContext();

UUPSUnsupportedProxiableUUID

Custom error with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4. solidity error UUPSUnsupportedProxiableUUID(bytes32 slot);

Upgrade

Event with signature Upgrade(address) and selector 0xf78721226efe9a1bb678189a16d1554928b9f2192e2cb93eeda83b79fa40007d. solidity event Upgrade(address implementation);

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b. solidity event Upgraded(address indexed implementation);

balancesCall

Function with signature balances(address) and selector 0x27e235e3. solidity function balances(address user) external view returns (uint256 amount);

balancesReturn

Container type for the return parameters of the balances(address) function.

constructorCall

Constructor`. solidity constructor();

depositCall

Function with signature deposit(address) and selector 0xf340fa01. solidity function deposit(address user) external payable;

depositReturn

Container type for the return parameters of the deposit(address) function.

getVersionCall

Function with signature getVersion() and selector 0x0d8e6e2c. solidity function getVersion() external pure returns (uint8 majorVersion, uint8 minorVersion, uint8 patchVersion);

getVersionReturn

Container type for the return parameters of the getVersion() function.

initializeCall

Function with signature initialize(address) and selector 0xc4d66de8. solidity function initialize(address multisig) external;

initializeReturn

Container type for the return parameters of the initialize(address) function.

maxDepositAmountCall

Function with signature maxDepositAmount() and selector 0x8ed83271. solidity function maxDepositAmount() external view returns (uint256);

maxDepositAmountReturn

Container type for the return parameters of the maxDepositAmount() function.

minDepositAmountCall

Function with signature minDepositAmount() and selector 0x645006ca. solidity function minDepositAmount() external view returns (uint256);

minDepositAmountReturn

Container type for the return parameters of the minDepositAmount() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b. solidity function owner() external view returns (address);

ownerReturn

Container type for the return parameters of the owner() function.

proxiableUUIDCall

Function with signature proxiableUUID() and selector 0x52d1902d. solidity function proxiableUUID() external view returns (bytes32);

proxiableUUIDReturn

Container type for the return parameters of the proxiableUUID() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6. solidity function renounceOwnership() external;

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b. solidity function transferOwnership(address newOwner) external;

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

upgradeToAndCallCall

Function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286. solidity function upgradeToAndCall(address newImplementation, bytes memory data) external payable;

upgradeToAndCallReturn

Container type for the return parameters of the upgradeToAndCall(address,bytes) function.

Enums

FeeContractCalls

Container for all the FeeContract function calls.

FeeContractErrors

Container for all the FeeContract custom errors.

FeeContractEvents

Container for all the FeeContract events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain FeeContract contract instance.