contract_bindings_alloy::lightclient::LightClient
Struct LightClientInstance

pub struct LightClientInstance<T, P, N = Ethereum> {
    address: Address,
    provider: P,
    _network_transport: PhantomData<(N, T)>,
}

A LightClient instance.

Contains type-safe methods for interacting with an on-chain instance of the
[`LightClient`](self) contract located at a given `address`, using a given
provider `P`.

If the contract bytecode is available (see the [`sol!`](alloy_sol_types::sol!)
documentation on how to provide it), the `deploy` and `deploy_builder` methods can
be used to deploy a new instance of the contract.

See the [module-level documentation](self) for all the available methods.

Fields
address: Addressprovider: P_network_transport: PhantomData<(N, T)>
Implementations

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> LightClientInstance<T, P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain LightClient contract instance.

    See the [wrapper's documentation](`LightClientInstance`) for more details.

pub async fn deploy(provider: P) -> Result<LightClientInstance<T, P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

    Returns a new instance of the contract, if the deployment was successful.

    For more fine-grained control over the deployment process, use [`deploy_builder`] instead.

pub fn deploy_builder(provider: P) -> RawCallBuilder<T, P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

    This is a simple wrapper around creating a `RawCallBuilder` with the data set to
    the bytecode concatenated with the constructor's ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<T, P: Clone, N> LightClientInstance<T, &P, N>

pub fn with_cloned_provider(self) -> LightClientInstance<T, P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> LightClientInstance<T, P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<T, &P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn UPGRADE_INTERFACE_VERSION( &self, ) -> SolCallBuilder<T, &P, UPGRADE_INTERFACE_VERSIONCall, N>

Creates a new call builder for the [UPGRADE_INTERFACE_VERSION] function.

pub fn currentBlockNumber( &self, ) -> SolCallBuilder<T, &P, currentBlockNumberCall, N>

Creates a new call builder for the [currentBlockNumber] function.

pub fn disablePermissionedProverMode( &self, ) -> SolCallBuilder<T, &P, disablePermissionedProverModeCall, N>

Creates a new call builder for the [disablePermissionedProverMode] function.

pub fn finalizedState(&self) -> SolCallBuilder<T, &P, finalizedStateCall, N>

Creates a new call builder for the [finalizedState] function.

pub fn genesisStakeTableState( &self, ) -> SolCallBuilder<T, &P, genesisStakeTableStateCall, N>

Creates a new call builder for the [genesisStakeTableState] function.

pub fn genesisState(&self) -> SolCallBuilder<T, &P, genesisStateCall, N>

Creates a new call builder for the [genesisState] function.

pub fn getHotShotCommitment( &self, hotShotBlockHeight: U256, ) -> SolCallBuilder<T, &P, getHotShotCommitmentCall, N>

Creates a new call builder for the [getHotShotCommitment] function.

pub fn getStateHistoryCount( &self, ) -> SolCallBuilder<T, &P, getStateHistoryCountCall, N>

Creates a new call builder for the [getStateHistoryCount] function.

pub fn getVersion(&self) -> SolCallBuilder<T, &P, getVersionCall, N>

Creates a new call builder for the [getVersion] function.

pub fn initialize( &self, _genesis: <LightClientState as SolType>::RustType, _genesisStakeTableState: <StakeTableState as SolType>::RustType, _stateHistoryRetentionPeriod: u32, owner: Address, ) -> SolCallBuilder<T, &P, initializeCall, N>

Creates a new call builder for the [initialize] function.

pub fn isPermissionedProverEnabled( &self, ) -> SolCallBuilder<T, &P, isPermissionedProverEnabledCall, N>

Creates a new call builder for the [isPermissionedProverEnabled] function.

pub fn lagOverEscapeHatchThreshold( &self, blockNumber: U256, blockThreshold: U256, ) -> SolCallBuilder<T, &P, lagOverEscapeHatchThresholdCall, N>

Creates a new call builder for the [lagOverEscapeHatchThreshold] function.

pub fn newFinalizedState( &self, newState: <LightClientState as SolType>::RustType, proof: <PlonkProof as SolType>::RustType, ) -> SolCallBuilder<T, &P, newFinalizedStateCall, N>

Creates a new call builder for the [newFinalizedState] function.

pub fn owner(&self) -> SolCallBuilder<T, &P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn permissionedProver( &self, ) -> SolCallBuilder<T, &P, permissionedProverCall, N>

Creates a new call builder for the [permissionedProver] function.

pub fn proxiableUUID(&self) -> SolCallBuilder<T, &P, proxiableUUIDCall, N>

Creates a new call builder for the [proxiableUUID] function.

pub fn renounceOwnership( &self, ) -> SolCallBuilder<T, &P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn setPermissionedProver( &self, prover: Address, ) -> SolCallBuilder<T, &P, setPermissionedProverCall, N>

Creates a new call builder for the [setPermissionedProver] function.

pub fn setstateHistoryRetentionPeriod( &self, historySeconds: u32, ) -> SolCallBuilder<T, &P, setstateHistoryRetentionPeriodCall, N>

Creates a new call builder for the [setstateHistoryRetentionPeriod] function.

pub fn stateHistoryCommitments( &self, _0: U256, ) -> SolCallBuilder<T, &P, stateHistoryCommitmentsCall, N>

Creates a new call builder for the [stateHistoryCommitments] function.

pub fn stateHistoryFirstIndex( &self, ) -> SolCallBuilder<T, &P, stateHistoryFirstIndexCall, N>

Creates a new call builder for the [stateHistoryFirstIndex] function.

pub fn stateHistoryRetentionPeriod( &self, ) -> SolCallBuilder<T, &P, stateHistoryRetentionPeriodCall, N>

Creates a new call builder for the [stateHistoryRetentionPeriod] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<T, &P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn upgradeToAndCall( &self, newImplementation: Address, data: Bytes, ) -> SolCallBuilder<T, &P, upgradeToAndCallCall, N>

Creates a new call builder for the [upgradeToAndCall] function.

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> LightClientInstance<T, P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<T, &P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn Initialized_filter(&self) -> Event<T, &P, Initialized, N>

Creates a new event filter for the Initialized event.

pub fn NewState_filter(&self) -> Event<T, &P, NewState, N>

Creates a new event filter for the NewState event.

pub fn OwnershipTransferred_filter( &self, ) -> Event<T, &P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.

pub fn PermissionedProverNotRequired_filter( &self, ) -> Event<T, &P, PermissionedProverNotRequired, N>

Creates a new event filter for the PermissionedProverNotRequired event.

pub fn PermissionedProverRequired_filter( &self, ) -> Event<T, &P, PermissionedProverRequired, N>

Creates a new event filter for the PermissionedProverRequired event.

pub fn Upgrade_filter(&self) -> Event<T, &P, Upgrade, N>

Creates a new event filter for the Upgrade event.

pub fn Upgraded_filter(&self) -> Event<T, &P, Upgraded, N>

Creates a new event filter for the Upgraded event.