contract_bindings_alloy::lightclientmock::LightClientMock
Enum LightClientMockCalls

pub enum LightClientMockCalls {

    UPGRADE_INTERFACE_VERSION(UPGRADE_INTERFACE_VERSIONCall),
    currentBlockNumber(currentBlockNumberCall),
    disablePermissionedProverMode(disablePermissionedProverModeCall),
    finalizedState(finalizedStateCall),
    genesisStakeTableState(genesisStakeTableStateCall),
    genesisState(genesisStateCall),
    getHotShotCommitment(getHotShotCommitmentCall),
    getStateHistoryCount(getStateHistoryCountCall),
    getVersion(getVersionCall),
    initialize(initializeCall),
    isPermissionedProverEnabled(isPermissionedProverEnabledCall),
    lagOverEscapeHatchThreshold(lagOverEscapeHatchThresholdCall),
    newFinalizedState(newFinalizedStateCall),
    owner(ownerCall),
    permissionedProver(permissionedProverCall),
    proxiableUUID(proxiableUUIDCall),
    renounceOwnership(renounceOwnershipCall),
    setFinalizedState(setFinalizedStateCall),
    setHotShotDownSince(setHotShotDownSinceCall),
    setHotShotUp(setHotShotUpCall),
    setPermissionedProver(setPermissionedProverCall),
    setStateHistory(setStateHistoryCall),
    setstateHistoryRetentionPeriod(setstateHistoryRetentionPeriodCall),
    stateHistoryCommitments(stateHistoryCommitmentsCall),
    stateHistoryFirstIndex(stateHistoryFirstIndexCall),
    stateHistoryRetentionPeriod(stateHistoryRetentionPeriodCall),
    transferOwnership(transferOwnershipCall),
    upgradeToAndCall(upgradeToAndCallCall),

}

Container for all the LightClientMock function calls.

Variants

UPGRADE_INTERFACE_VERSION(UPGRADE_INTERFACE_VERSIONCall)

currentBlockNumber(currentBlockNumberCall)

disablePermissionedProverMode(disablePermissionedProverModeCall)

finalizedState(finalizedStateCall)

genesisStakeTableState(genesisStakeTableStateCall)

genesisState(genesisStateCall)

getHotShotCommitment(getHotShotCommitmentCall)

getStateHistoryCount(getStateHistoryCountCall)

getVersion(getVersionCall)

initialize(initializeCall)

isPermissionedProverEnabled(isPermissionedProverEnabledCall)

lagOverEscapeHatchThreshold(lagOverEscapeHatchThresholdCall)

newFinalizedState(newFinalizedStateCall)

owner(ownerCall)

permissionedProver(permissionedProverCall)

proxiableUUID(proxiableUUIDCall)

renounceOwnership(renounceOwnershipCall)

setFinalizedState(setFinalizedStateCall)

setHotShotDownSince(setHotShotDownSinceCall)

setHotShotUp(setHotShotUpCall)

setPermissionedProver(setPermissionedProverCall)

setStateHistory(setStateHistoryCall)

setstateHistoryRetentionPeriod(setstateHistoryRetentionPeriodCall)

stateHistoryCommitments(stateHistoryCommitmentsCall)

stateHistoryFirstIndex(stateHistoryFirstIndexCall)

stateHistoryRetentionPeriod(stateHistoryRetentionPeriodCall)

transferOwnership(transferOwnershipCall)

upgradeToAndCall(upgradeToAndCallCall)

Implementations

impl LightClientMockCalls

pub const SELECTORS: &'static [[u8; 4]] = _

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.