contract_bindings_alloy::permissionedstaketable
Module PermissionedStakeTable

Generated by the following Solidity interface…

library BN254 {
    type BaseField is uint256;
    struct G2Point {
        BaseField x0;
        BaseField x1;
        BaseField y0;
        BaseField y1;
    }
}

library EdOnBN254 {
    struct EdOnBN254Point {
        uint256 x;
        uint256 y;
    }
}

interface PermissionedStakeTable {
    struct NodeInfo {
        BN254.G2Point blsVK;
        EdOnBN254.EdOnBN254Point schnorrVK;
        bool isDA;
    }

    error InvalidInitialization();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error StakerAlreadyExists(BN254.G2Point);
    error StakerNotFound(BN254.G2Point);

    event Initialized(uint64 version);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event StakersUpdated(BN254.G2Point[] removed, NodeInfo[] added);

    constructor(NodeInfo[] initialStakers);

    function _hashBlsKey(BN254.G2Point memory blsVK) external pure returns (bytes32);
    function initialize() external;
    function initializedAtBlock() external view returns (uint256);
    function isStaker(BN254.G2Point memory staker) external view returns (bool);
    function owner() external view returns (address);
    function renounceOwnership() external;
    function transferOwnership(address newOwner) external;
    function update(BN254.G2Point[] memory stakersToRemove, NodeInfo[] memory newStakers) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "initialStakers",
        "type": "tuple[]",
        "internalType": "struct PermissionedStakeTable.NodeInfo[]",
        "components": [
          {
            "name": "blsVK",
            "type": "tuple",
            "internalType": "struct BN254.G2Point",
            "components": [
              {
                "name": "x0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "x1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              }
            ]
          },
          {
            "name": "schnorrVK",
            "type": "tuple",
            "internalType": "struct EdOnBN254.EdOnBN254Point",
            "components": [
              {
                "name": "x",
                "type": "uint256",
                "internalType": "uint256"
              },
              {
                "name": "y",
                "type": "uint256",
                "internalType": "uint256"
              }
            ]
          },
          {
            "name": "isDA",
            "type": "bool",
            "internalType": "bool"
          }
        ]
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "_hashBlsKey",
    "inputs": [
      {
        "name": "blsVK",
        "type": "tuple",
        "internalType": "struct BN254.G2Point",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initializedAtBlock",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isStaker",
    "inputs": [
      {
        "name": "staker",
        "type": "tuple",
        "internalType": "struct BN254.G2Point",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "update",
    "inputs": [
      {
        "name": "stakersToRemove",
        "type": "tuple[]",
        "internalType": "struct BN254.G2Point[]",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      },
      {
        "name": "newStakers",
        "type": "tuple[]",
        "internalType": "struct PermissionedStakeTable.NodeInfo[]",
        "components": [
          {
            "name": "blsVK",
            "type": "tuple",
            "internalType": "struct BN254.G2Point",
            "components": [
              {
                "name": "x0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "x1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              }
            ]
          },
          {
            "name": "schnorrVK",
            "type": "tuple",
            "internalType": "struct EdOnBN254.EdOnBN254Point",
            "components": [
              {
                "name": "x",
                "type": "uint256",
                "internalType": "uint256"
              },
              {
                "name": "y",
                "type": "uint256",
                "internalType": "uint256"
              }
            ]
          },
          {
            "name": "isDA",
            "type": "bool",
            "internalType": "bool"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "StakersUpdated",
    "inputs": [
      {
        "name": "removed",
        "type": "tuple[]",
        "indexed": false,
        "internalType": "struct BN254.G2Point[]",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      },
      {
        "name": "added",
        "type": "tuple[]",
        "indexed": false,
        "internalType": "struct PermissionedStakeTable.NodeInfo[]",
        "components": [
          {
            "name": "blsVK",
            "type": "tuple",
            "internalType": "struct BN254.G2Point",
            "components": [
              {
                "name": "x0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "x1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y0",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              },
              {
                "name": "y1",
                "type": "uint256",
                "internalType": "BN254.BaseField"
              }
            ]
          },
          {
            "name": "schnorrVK",
            "type": "tuple",
            "internalType": "struct EdOnBN254.EdOnBN254Point",
            "components": [
              {
                "name": "x",
                "type": "uint256",
                "internalType": "uint256"
              },
              {
                "name": "y",
                "type": "uint256",
                "internalType": "uint256"
              }
            ]
          },
          {
            "name": "isDA",
            "type": "bool",
            "internalType": "bool"
          }
        ]
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "StakerAlreadyExists",
    "inputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct BN254.G2Point",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      }
    ]
  },
  {
    "type": "error",
    "name": "StakerNotFound",
    "inputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct BN254.G2Point",
        "components": [
          {
            "name": "x0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "x1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y0",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          },
          {
            "name": "y1",
            "type": "uint256",
            "internalType": "BN254.BaseField"
          }
        ]
      }
    ]
  }
]

Structs

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2. solidity event Initialized(uint64 version);

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9. solidity error InvalidInitialization();

NodeInfo

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8. solidity error NotInitializing();

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7. solidity error OwnableInvalidOwner(address owner);

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7. solidity error OwnableUnauthorizedAccount(address account);

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0. solidity event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);

PermissionedStakeTableInstance

A PermissionedStakeTable instance.

StakerAlreadyExists

Custom error with signature StakerAlreadyExists((uint256,uint256,uint256,uint256)) and selector 0x360dc282. solidity error StakerAlreadyExists(BN254.G2Point);

StakerNotFound

Custom error with signature StakerNotFound((uint256,uint256,uint256,uint256)) and selector 0x34a7561f. solidity error StakerNotFound(BN254.G2Point);

StakersUpdated

Event with signature StakersUpdated((uint256,uint256,uint256,uint256)[],((uint256,uint256,uint256,uint256),(uint256,uint256),bool)[]) and selector 0x4bb31cd9ae87a3f8289558c79be5232506ce36de4269c8ba4534a6fbe9e2965d. solidity event StakersUpdated(BN254.G2Point[] removed, NodeInfo[] added);

_hashBlsKeyCall

Function with signature _hashBlsKey((uint256,uint256,uint256,uint256)) and selector 0x9b30a5e6. solidity function _hashBlsKey(BN254.G2Point memory blsVK) external pure returns (bytes32);

_hashBlsKeyReturn

Container type for the return parameters of the _hashBlsKey((uint256,uint256,uint256,uint256)) function.

constructorCall

Constructor`. solidity constructor(NodeInfo[] initialStakers);

initializeCall

Function with signature initialize() and selector 0x8129fc1c. solidity function initialize() external;

initializeReturn

Container type for the return parameters of the initialize() function.

initializedAtBlockCall

Function with signature initializedAtBlock() and selector 0x3e9df9b5. solidity function initializedAtBlock() external view returns (uint256);

initializedAtBlockReturn

Container type for the return parameters of the initializedAtBlock() function.

isStakerCall

Function with signature isStaker((uint256,uint256,uint256,uint256)) and selector 0x75d705e9. solidity function isStaker(BN254.G2Point memory staker) external view returns (bool);

isStakerReturn

Container type for the return parameters of the isStaker((uint256,uint256,uint256,uint256)) function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b. solidity function owner() external view returns (address);

ownerReturn

Container type for the return parameters of the owner() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6. solidity function renounceOwnership() external;

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b. solidity function transferOwnership(address newOwner) external;

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

updateCall

Function with signature update((uint256,uint256,uint256,uint256)[],((uint256,uint256,uint256,uint256),(uint256,uint256),bool)[]) and selector 0xcbba7c78. solidity function update(BN254.G2Point[] memory stakersToRemove, NodeInfo[] memory newStakers) external;

updateReturn

Container type for the return parameters of the update((uint256,uint256,uint256,uint256)[],((uint256,uint256,uint256,uint256),(uint256,uint256),bool)[]) function.

Enums

PermissionedStakeTableCalls

Container for all the PermissionedStakeTable function calls.

PermissionedStakeTableErrors

Container for all the PermissionedStakeTable custom errors.

PermissionedStakeTableEvents

Container for all the PermissionedStakeTable events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain PermissionedStakeTable contract instance.